requirements.txt
opencv-python
pandas
streamlit
Pillow
numpy